+++ test/oli/delivery/submission_test.exs
@@ -97,7 +97,7 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
       assert Enum.empty?(user1_page_context.resource_attempts)
 
       # Start the attempt and go into the assessment
-      activity_provider = &Oli.Delivery.ActivityProvider.provide/2
+      activity_provider = &Oli.Delivery.ActivityProvider.provide/3
 
       {:ok,
        %Oli.Delivery.Attempts.PageLifecycle.AttemptState{
@@ -368,7 +368,7 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
       user1_activity_attempt1: activity_attempt,
       graded_page_user1_attempt1: resource_attempt1
     } do
-      activity_provider = &Oli.Delivery.ActivityProvider.provide/2
+      activity_provider = &Oli.Delivery.ActivityProvider.provide/3
 
       # User1 has a started resource attempt, so it should be "in progress"
       {:ok, {:in_progress, _resource_attempt}} =
@@ -409,7 +409,7 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
       user1_part1_attempt1: part_attempt,
       user1_activity_attempt1: activity_attempt
     } do
-      activity_provider = &Oli.Delivery.ActivityProvider.provide/2
+      activity_provider = &Oli.Delivery.ActivityProvider.provide/3
 
       {:ok, {:in_progress, _resource_attempt}} =
         PageLifecycle.visit(
@@ -449,7 +449,7 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
            graded_page_user2_attempt1: user2_resource_attempt1,
            user2: user2
          } do
-      activity_provider = &Oli.Delivery.ActivityProvider.provide/2
+      activity_provider = &Oli.Delivery.ActivityProvider.provide/3
 
       # User 1
       {:ok, {:in_progress, resource_attempt_user1}} =